+++ src/main/java/subway/exception/common/ExceptionHandlerController.java
@@ -1,4 +1,4 @@
-package subway.global.exception.common;
+package subway.exception.common;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
