+++ src/main/java/subway/service/SectionCommandService.java
@@ -7,7 +7,7 @@ import subway.dao.SectionEntity;
 import subway.domain.section.Section;
 import subway.domain.station.Station;
 import subway.domain.station.Stations;
-import subway.global.exception.section.CanNotDuplicatedSectionException;
+import subway.exception.section.CanNotDuplicatedSectionException;
 
 import java.util.List;
